refactor: update import paths for staking slices
